feat(users): add archive and restore functionality with confirmation modals to users[id] page from users page
